# places_app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doc API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/places": {
            "get": {
                "tags": [
                    "Place"
                ],
                "summary": "Список мест",
                "operationId": "2418de94a22e07adbd6a7d9ac162621e",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "14"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Some place"
                                                    },
                                                    "longitude": {
                                                        "type": "number",
                                                        "example": "14.324987"
                                                    },
                                                    "latitude": {
                                                        "type": "number",
                                                        "example": "-60.324987"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Place"
                ],
                "summary": "Создание записи о месте",
                "operationId": "9296af3eb6275aed1f9da38c38e76cf9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Some place"
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "maximum": 180,
                                                "minimum": -180,
                                                "example": 14.324987
                                            },
                                            "latitude": {
                                                "type": "number",
                                                "maximum": 90,
                                                "minimum": -90,
                                                "example": -60.324987
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "14"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Some place"
                                                },
                                                "longitude": {
                                                    "type": "number",
                                                    "example": "14.324987"
                                                },
                                                "latitude": {
                                                    "type": "number",
                                                    "example": "-60.324987"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/places/{place}": {
            "get": {
                "tags": [
                    "Place"
                ],
                "summary": "Запись о месте",
                "operationId": "a0631087447ce542ac9a4352aeb50c18",
                "parameters": [
                    {
                        "name": "place",
                        "in": "path",
                        "description": "ID места",
                        "required": true,
                        "example": 14
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "14"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Some place"
                                                },
                                                "longitude": {
                                                    "type": "number",
                                                    "example": "14.324987"
                                                },
                                                "latitude": {
                                                    "type": "number",
                                                    "example": "-60.324987"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/places/{place}/add": {
            "post": {
                "tags": [
                    "Place"
                ],
                "summary": "Добавление места в список желаемых для текущего пользователя",
                "operationId": "0846dd99462d79eef64db279ce812758",
                "parameters": [
                    {
                        "name": "place",
                        "in": "path",
                        "description": "ID места",
                        "required": true,
                        "example": 14
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "14"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Some place"
                                                },
                                                "longitude": {
                                                    "type": "number",
                                                    "example": "14.324987"
                                                },
                                                "latitude": {
                                                    "type": "number",
                                                    "example": "-60.324987"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "No more items can be added | This item has already been added",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "422"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No more items can be added"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/places/{place}/remove": {
            "delete": {
                "tags": [
                    "Place"
                ],
                "summary": "Удаление места из списка желаемых для текущего пользователя",
                "operationId": "231c066ba4a71d6483b1afb37172aeea",
                "parameters": [
                    {
                        "name": "place",
                        "in": "path",
                        "description": "ID места",
                        "required": true,
                        "example": 14
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "14"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Some place"
                                                },
                                                "longitude": {
                                                    "type": "number",
                                                    "example": "14.324987"
                                                },
                                                "latitude": {
                                                    "type": "number",
                                                    "example": "-60.324987"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "This item is not included in the list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": "422"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "This item is not included in the list"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}